ingest-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
python-dateutil==2.8.2
protobuf==4.25.0
opentelemetry-proto==1.21.0
httpx>=0.25.0
aiohttp==3.9.5
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
